Adds the primary support for Flexible Data Placement(FDP) over NVMe into Cachelib
Summary:
This commit adds the device layer support for NVMe-FDP semantics and adds the RUH-awareness feature of NVMe-FDP in
the upper layers of Navy. This allows the BlockCache(large items) and BigHash(small items) of Navy to segregate their data streams
in physical NAND media by using the FDP placement Identifiers.

With this changes, the Cachelib can reduce the Device Write Amplification (WAF) significantly even in high SSD utilization scenarios
("nvmCacheSizeMB" above 50% of the SSD capacity) in most of the cachelib workloads.

This commit introduces a 'placementHandle' concept for data placement, which can be used by both BC and BH of Navy on
device write() calls, especially for FDP placements. The 'placementHandle' have to be allocated from the device.

io_uring_cmd interface(through nvme char device) is used to send FDP directives to Linux kernel, as sending it through the
conventional block interfaces is not suported yet. The user needs to enable fdpMode and configure an NVMe char device instead
of the regular block device to enable the FDP mode. (Ex:"fdpMode": true, "nvmCachePaths": ["/dev/ng0n1"])

Signed-off-by: Arun George <arun.george@samsung.com>