Adding Unit Tests for FdpNvme class
summary:
This commit adds the unit test cases for "FdpNvme" class. It also adds a new constructor for FdpNvme.
The existing constructor of FdpNvme communicates with the actual FDP device and makes the unit tests
impossible to run without the actual device. So this commit adds a new constructor which allows to
experiment with "FdpNvme" without the actual FDP device.

Signed-off-by: Shrihari E S  <shrihari.s@samsung.com>